Added Context, Checking for duplicates
