+++ src/index.js
@@ -4,11 +4,14 @@ import { FirestoreProvider } from 'react-firestore';
 import { fb as firebase } from './lib/firebase.js';
 import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
+import { ListContextProvider } from './listContext';
 
 ReactDOM.render(
   <FirestoreProvider firebase={firebase}>
     <Router>
-      <App />
+      <ListContextProvider>
+        <App />
+      </ListContextProvider>
     </Router>
   </FirestoreProvider>,
   document.getElementById('root'),
